We had some issues with the driver and hibernate in old versions of the driver. (Note: we are also on quite old version of hibernate: 4.3.7) We noted that hibernate was re-setting the isolation level when it reuses a connection from the connection pool
We were *not* able to simulate this with the Postgresql-42.2.20 driver.

It was decided to *still apply the previous patch* in order to play it safe.

With this patch, if the driver is required to set the transaction isolation level to the same level that the transaction
is already running with, the request is simply ignored, instead of throwing an exception.
Unfortunately the patch incurs a performance hit, because it runs the query "SHOW TRANSACTION ISOLATION LEVEL" in order
to compare the current level with the requested one.